+++ crates/lib/src/games/definitions.rs
@@ -136,5 +136,6 @@ pub static GAMES: Map<&'static str, Game> = phf_map! {
     "unrealtournament2004" => game!("Unreal Tournament 2004", 7778, Protocol::Unreal2),
     "eco" => game!("Eco", 3000, Protocol::PROPRIETARY(ProprietaryProtocol::Eco)),
     "zps" => game!("Zombie Panic: Source", 27015, Protocol::Valve(Engine::new(17_500))),
+    "mordhau" => game!("Mordhau", 27015, Protocol::Valve(Engine::new(629_760))),
     "mindustry" => game!("Mindustry", crate::games::mindustry::DEFAULT_PORT, Protocol::PROPRIETARY(ProprietaryProtocol::Mindustry)),
 };