+++ crates/lib/src/games/definitions.rs
@@ -82,6 +82,7 @@ pub static GAMES: Map<&'static str, Game> = phf_map! {
     "dods" => game!("Day of Defeat: Source", 27015, Protocol::Valve(Engine::new(300))),
     "doi" => game!("Day of Infamy", 27015, Protocol::Valve(Engine::new(447_820))),
     "dst" => game!("Don't Starve Together", 27016, Protocol::Valve(Engine::new(322_320))),
+    "enshrouded" => game!("Enshrouded", 15637, Protocol::Valve(Engine::new(1_203_620))),
     "ffow" => game!("Frontlines: Fuel of War", 5478, Protocol::PROPRIETARY(ProprietaryProtocol::FFOW)),
     "garrysmod" => game!("Garry's Mod", 27016, Protocol::Valve(Engine::new(4000))),
     "hl2d" => game!("Half-Life 2 Deathmatch", 27015, Protocol::Valve(Engine::new(320))),