+++ crates/lib/src/protocols/mod.rs
@@ -4,6 +4,9 @@
 //! implementation will be in that specific needed place, a protocol can be
 //! independently queried.
 
+#[cfg(feature = "tls")]
+/// Reference: [node-GameDig](https://github.com/gamedig/node-gamedig/blob/master/protocols/epic.js)
+pub mod epic;
 /// Reference: [node-GameDig](https://github.com/gamedig/node-gamedig/blob/master/protocols/gamespy1.js)
 pub mod gamespy;
 /// Reference: [node-GameDig](https://github.com/gamedig/node-gamedig/blob/master/protocols/quake1.js)