+++ crates/lib/src/protocols/types.rs
@@ -1,5 +1,7 @@
 #[cfg(feature = "games")]
 use crate::games::minecraft;
+#[cfg(feature = "tls")]
+use crate::protocols::epic;
 use crate::protocols::{gamespy, quake, unreal2, valve};
 use crate::GDErrorKind::InvalidInput;
 use crate::GDResult;
@@ -31,6 +33,8 @@ pub enum Protocol {
     Quake(quake::QuakeVersion),
     Valve(valve::Engine),
     Unreal2,
+    #[cfg(feature = "tls")]
+    Epic(epic::Credentials),
     #[cfg(feature = "games")]
     PROPRIETARY(ProprietaryProtocol),
 }
@@ -43,6 +47,8 @@ pub enum GenericResponse<'a> {
     Quake(quake::VersionedResponse<'a>),
     Valve(&'a valve::Response),
     Unreal2(&'a unreal2::Response),
+    #[cfg(feature = "tls")]
+    Epic(&'a epic::Response),
     #[cfg(feature = "games")]
     Mindustry(&'a crate::games::mindustry::types::ServerData),
     #[cfg(feature = "games")]
@@ -68,6 +74,8 @@ pub enum GenericPlayer<'a> {
     QuakeTwo(&'a quake::two::Player),
     Gamespy(gamespy::VersionedPlayer<'a>),
     Unreal2(&'a unreal2::Player),
+    #[cfg(feature = "tls")]
+    Epic(&'a epic::Player),
     #[cfg(feature = "games")]
     Minecraft(&'a minecraft::Player),
     #[cfg(feature = "games")]