+++ crates/lib/src/games/definitions.rs
@@ -151,4 +151,5 @@ pub static GAMES: Map<&'static str, Game> = phf_map! {
     "moe" => game!("Myth Of Empires", 12888, Protocol::Valve(Engine::new(1_371_580))),
     "mordhau" => game!("Mordhau", 27015, Protocol::Valve(Engine::new(629_760))),
     "mindustry" => game!("Mindustry", crate::games::mindustry::DEFAULT_PORT, Protocol::PROPRIETARY(ProprietaryProtocol::Mindustry)),
+    "nla" => game!("Nova-Life: Amboise", 27015, Protocol::Valve(Engine::new(885_570))),
 };