+++ crates/lib/CHANGELOG.md
@@ -2,7 +2,8 @@
 Who knows what the future holds...
 
 # 0.X.Y - DD/MM/YYYY
-Nothing... yet!
+Games:
+- [Mordhau](https://store.steampowered.com/app/629760/MORDHAU/) support.
 
 # 0.5.0 - 15/03/2024
 ### Changes: