+++ crates/lib/CHANGELOG.md
@@ -5,6 +5,7 @@ Who knows what the future holds...
 Games:
 - [Mordhau](https://store.steampowered.com/app/629760/MORDHAU/) support.
 - [Enshrouded](https://store.steampowered.com/app/1203620/Enshrouded/) support.
+- [Myth of Empires](https://store.steampowered.com/app/1371580/Myth_of_Empires/) support.
 
 # 0.5.0 - 15/03/2024
 ### Changes: