+++ crates/lib/CHANGELOG.md
@@ -2,6 +2,10 @@ Who knows what the future holds...
 
 # 0.X.Y - DD/MM/YYYY
 
+To be made...
+
+# 0.5.1 - 12/05/2024
+
 Games:
 
 - [Mordhau](https://store.steampowered.com/app/629760/MORDHAU/) support.